pyproject.toml
#################################################
# Core
#################################################

[project]
name = "pipeline"
version = "0.1.0"
description = "Simple RESTful API for pipeline configuration"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "click>=8.1.7",
    "fastapi>=0.115.2",
    "gunicorn>=23.0.0",
    "pydantic-settings>=2.5.2",
    "uvicorn>=0.32.0",
]

[tool.uv]
dev-dependencies = [
    "httpx>=0.27.2",
    "mypy>=1.12.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "pytest-order>=1.3.0",
    "pytest>=8.3.3",
    "ruff>=0.6.9",
]

#################################################
# Ruff
#################################################

[tool.ruff]
respect-gitignore = true
preview = true
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "PT", "C4", "I", "D", "LOG", "G", "T20", "FAST"]
ignore = ["E501", "D102"]

# Now, create a specific section for the tests directory
[tool.ruff.lint.per-file-ignores]
"tests/*" = ["D"]

[tool.ruff.lint.mccabe]
max-complexity = 10

[tool.ruff.lint.pydocstyle]
convention = "google"

#################################################
# Pytest
#################################################

[tool.pytest.ini_options]
addopts = "--strict-markers --maxfail 1 --cov tests/ --no-header"
markers = """
    integration: mark a test as an integration test.
"""
console_output_style = "progress"
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
